+++ src/components/link-preview/video.jsx
@@ -11,7 +11,7 @@ import * as aspectRatio from './helpers/size-cache';
 
 const YOUTUBE_VIDEO_RE =
   /^https?:\/\/(?:www\.|m\.)?(?:youtu\.be\/|youtube\.com\/(?:embed\/|shorts\/|v\/|watch\?(?:v=|.+&v=)))([\w-]+)/i;
-const VIMEO_VIDEO_RE = /^https?:\/\/vimeo\.com\/(\d+)(?:\/([a-z\d]+))?/i;
+const VIMEO_VIDEO_RE = /^https?:\/\/vimeo\.com\/(?:ondemand\/[^/]+\/)?(\d+)(?:\/([a-z\d]+))?/i;
 const COUB_VIDEO_RE = /^https?:\/\/coub\.com\/view\/([a-z\d]+)/i;
 const IMGUR_VIDEO_RE = /^https?:\/\/i\.imgur\.com\/([a-z\d]+)\.(gifv|mp4)/i;
 const GFYCAT_RE =