+++ src/components/user-profile.jsx
@@ -33,7 +33,6 @@ export default function UserProfile(props) {
 
         {props.canIPostHere && (
           <CreatePost
-            createPostViewState={props.createPostViewState}
             sendTo={props.sendTo}
             user={props.user}
             createPost={props.createPost}